renaming browsing.js to visiting.js
